backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.27.0
gunicorn>=21.0.0

sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiosqlite>=0.19.0
alembic>=1.13.0

python-jose[cryptography]>=3.3.0

